RM 2046: fixed bugs causing AMICA images to be all-black.
The problem was exposed, but unfortunately not detected,
by changes made for #1283. Specifically, as part of
changes for that issue, some logic was simplified
based on the assumption that an image file was one type
or other of a file: either a FITS file, or a PNG
file (or ENVI file), but not both, i.e., an image would
not be associated with both a FITS and a PNG file. This
is clearly the way the code is intended to function
and mostly does function.

However, there was an undetected bug (probably originating
in copy-and-paste) which resulted in a violation of
the assumption. If an image used a LABEL file, both the
FITS file name AND the PNG file name were being set
(with the erroneous "PNG" file name being the same name
as the label file name!) The simplified logic
from #1283 tested only for whether the PNG file name was
set, and in this case it initialized the image range
using the VTK image range. Unfortunately, the latter
range was from the FITS file, which had fill values of
1.e38 in it, leading to the all-black appearance.

Three small changes were made to fix the problem.
1) The bug was corrected so that only if the file really
is a PNG file will the PNG file name be set.
2) An unrelated bug that was harmless was also fixed in
the AMICA image class, wherein it did not correctly apply
flips and rotations to the image after doing the flat
field correction. This is a moot point in the AMICA case,
since it has no flips and rotations, but on principle,
this was changed to make it consistent with the rest of the
code. This bug was recently introduced as part of unrelated
image code refactoring.
3) Out of an abundance of caution, the simplified logic
was modified to trigger the PNG specific code only if
the PNG file is set AND the FITS file is null, i.e., it
again behaves the way it did before the change made in
#1283. In all likelihood, the whole block of code could
be removed, since it's not clear why it's there or if it's
ever being executed, but I'm not eager to take a chance on that.